% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BV_LDER_GE_mainFUNs.R
\name{runBV_LDER_GE}
\alias{runBV_LDER_GE}
\title{Main function}
\usage{
runBV_LDER_GE(
  assoc_gwis,
  assoc_gwas,
  n.gwis,
  n.gwas,
  n.overlap,
  path,
  LD.insample = T,
  n.ld,
  method = "lder",
  type = "jack",
  size_num = 200
)
}
\arguments{
\item{n.gwas}{The sample size of the GWAS summary statistics}

\item{path}{The path of LD panel directory}

\item{LD.insample}{T/F, whether the LD reference is estimated with the GWAS cohort (T) or external reference panel (e.g. 1000 Genome Project) (F)}

\item{n.ld}{The sample size of the LD reference}

\item{method}{'lder', 'ldsc', or 'both'}

\item{type}{if 'jack' then conducts delete-wise block jackknife. if 'none' then does not do inference}

\item{size_num}{Number of blocks for jackknife}

\item{assoc}{GWIS summary statistics, need to include snp, chr, a0, a1, z (header is necessary)}

\item{cores}{The number of cores for computation in parallel}
}
\description{
Run LDER-GE
}
