% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BV_LDER_GE_adjFUNs.R
\name{runBV_LDER_GE_adj}
\alias{runBV_LDER_GE_adj}
\title{Main function}
\usage{
runBV_LDER_GE_adj(
  assoc_gwis_Y,
  assoc_gwas_Y,
  assoc_gwas_E,
  n.gwis_Y,
  n.gwas_Y,
  n.gwas_E,
  path,
  LD.insample = T,
  n.ld,
  method = "lder",
  size_num = 200,
  R2 = 0
)
}
\arguments{
\item{assoc_gwis_Y}{GWIS (GE interaction effect) summary statistics for Y, need to include snp, chr, a0, a1, z (header is necessary)}

\item{assoc_gwas_Y}{GWAS (additive genetic effect) summary statistics for Y, need to include snp, chr, a0, a1, z (header is necessary)}

\item{assoc_gwas_E}{GWAS (additive genetic effect) summary statistics for E, need to include snp, chr, a0, a1, z (header is necessary)}

\item{n.gwis_Y}{The sample size of the GWIS (GE interaction effect) summary statistics for Y}

\item{n.gwas_Y}{The sample size of the GWAS (additive genetic effect) summary statistics for Y}

\item{n.gwas_E}{The sample size of the GWAS (additive genetic effect) summary statistics for E}

\item{path}{The path of LD panel directory}

\item{LD.insample}{T/F, whether the LD reference is estimated with target cohort (T) or external reference panel (e.g. 1000 Genome Project) (F)}

\item{n.ld}{The sample size of the LD reference}

\item{method}{'lder', 'ldsc', or 'both'}

\item{size_num}{Number of blocks for jackknife}

\item{R2}{R2 between E and Y. If focusing on conditional heritability (i.e., Y adjusted for E already) then set to 0.}
}
\description{
Run BVLDER-GE-adjust
}
