% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BV_LDER_GE_mainFUNs.R
\name{runBV_LDER_GE}
\alias{runBV_LDER_GE}
\title{Main function}
\usage{
runBV_LDER_GE(
  assoc_gwis,
  assoc_gwas,
  n.gwis,
  n.gwas,
  n.overlap = NULL,
  path,
  LD.insample = T,
  n.ld,
  method = "lder",
  size_num = 200
)
}
\arguments{
\item{assoc_gwis}{GWIS (GE interaction effect) summary statistics, need to include snp, chr, a0, a1, z (header is necessary)}

\item{assoc_gwas}{GWAS (additive genetic effect) summary statistics, need to include snp, chr, a0, a1, z (header is necessary)}

\item{n.gwis}{The sample size of the GWIS (GE interaction effect) summary statistics}

\item{n.gwas}{The sample size of the GWAS (additive genetic effect) summary statistics}

\item{path}{The path of LD panel directory}

\item{LD.insample}{T/F, whether the LD reference is estimated with target cohort (T) or external reference panel (e.g. 1000 Genome Project) (F)}

\item{n.ld}{The sample size of the LD reference panel}

\item{method}{'lder', 'ldsc', or 'both'}

\item{size_num}{Number of blocks for jackknife}
}
\description{
Run BVLDER-GE
}
